Safari-611.1.10/LayoutTests/fast/encoding/legacy-utf-32-aliases.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("This test verifies that certain aliases for UTF-32 are not supported.");

debug("The following encodings should not be supported");

let utf32LegacyAliases = [
"ISO-10646-UCS-4", "ibm-1236", "ibm-1237", "csUCS4", "ucs-4", "UTF32_LittleEndian", "ibm-1234", "ibm-1235", "UTF32_BigEndian", "ibm-1232", "ibm-1233", "ibm-9424"
]; 

for (let encoding of utf32LegacyAliases) {
    let canonical_name_expr = 'new TextDecoder("' + encoding + '").encoding';
    shouldThrow(canonical_name_expr);
}

</script>
</body>
</html>
